move text processing fn to new file
